requirements.txt
Pillow==9.2.0
pip==22.2.2
wheel==0.37.1
pytz==2022.2.1
numpy==1.23.3
setuptools==65.3.0
pandas==1.5.0
python-dateutil==2.8.2
six==1.16.0
opencv-python==4.6.0.66
pyzbar==0.1.9